webapi/webapi-webrtc-w3c-tests/webrtc/RTCDataChannel.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang,Hongjuan <hongjuanx.wang@intel.com>

-->

<html>
  <head>
    <meta charset='utf-8'>
    <title>WebRTC Test: RTCDataChannel</title>
    <link rel='author' title='Intel' href='http://www.intel.com'>
    <link rel='help' href='http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface'>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="./support/support.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>

      var pc;

      setup(function () {
        if (typeof RTCPeerConnection != "undefined") {
          pc = new RTCPeerConnection(null,null);
        } else {
          pc = new webkitRTCPeerConnection(null,null);
        }
        RTCDataChannel = pc.createDataChannel("label");
      });

      test(function () {
        assert_equals(RTCDataChannel.binaryType, "blob", "the binaryType attribute init value");
      }, "Check if the setting of RTCDataChannel.binaryType is normally");

      test(function () {
        assert_equals(RTCDataChannel.bufferedAmount, 0, "the RTCDataChannel.bufferedAmount init value");
      }, "Check if the setting of RTCDataChannel.bufferedAmount is normally");

      test(function () {
        assert_not_equals(RTCDataChannel.id, null, "the RTCDataChannel.id init value");
      }, "Check if the setting of RTCDataChannel.id is normally");

      test(function () {
        assert_not_equals(RTCDataChannel.label, null, "the RTCDataChannel.label init value");
      }, "Check if the setting of RTCDataChannel.label is normally");

      test(function () {
        assert_equals(RTCDataChannel.maxRetransmits, 65535, "the RTCDataChannel.maxRetransmits init value");
      }, "Check if the setting of RTCDataChannel.maxRetransmits is normally");

      test(function () {
        assert_not_equals(RTCDataChannel.maxRetransmitTime, null, "the RTCDataChannel.maxRetransmitTime init value");
      }, "Check if the setting of RTCDataChannel.maxRetransmitTime is normally");

      test(function () {
        assert_equals(RTCDataChannel.negotiated, false, "the RTCDataChannel.negotiated init value");
      }, "Check if the setting of RTCDataChannel.negotiated is normally");

      test(function () {
        assert_true(RTCDataChannel.ordered, "the RTCDataChannel.ordered init value");
      }, "Check if the setting of RTCDataChannel.ordered is normally");

      test(function () {
        assert_equals(RTCDataChannel.protocol, "", "the RTCDataChannel.protocol init value");
      }, "Check if the setting of RTCDataChannel.protocol is normally");

      test(function () {
        assert_equals(RTCDataChannel.readyState, "connecting", "the RTCDataChannel.readyState init value");
      }, "Check if the setting of RTCDataChannel.readyState is normally");
    </script>
  </body>
</html>
